+++ umdb-app-client/src/components/Footer.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+const Footer = () => {
+    return (
+        <footer style={{ padding: '20px 0px' }}>
+            <div className="container">
+                <p style={{ color: '#a1a1a1', textAlign: 'center' }}>© Copyright 2020 - UMDb® - Todos os direitos reservados.</p>
+            </div>
+        </footer>
+    );
+}
+
+export default Footer;
