--- umdb-app-client/src/components/Header/Header.css
@@ -1,3 +1,0 @@
-.menu .item:hover {
-    background-color: yellowgreen;
-}
